Initial Face-Recognition Module
I looked up some basics on how to detect faces with OpenCV in python and created a basic module that does just that.
This module works by using various built-in methods provided by OpenCV, including a model trained to detect faces. I will add explanations as I understand how it works myself.

This module also includes testing capabilities for a single image; just run the script in the command line with an argument specifying the path to an image with a face in it. (Example: `python face_recognition.py path\to\image.png`

Next step will be to hook it up to video input to better test its capabilities, then look into how we could go about detecting the distance from the face to the camera. (I believe the 'scaleFactor' field for the method that actually detects faces may help us in that, but I'm not quite sure how the facial recognition works under the hood quite yet)