Added some checks to prevent errors with camera capture
Previously, the code assumed that a camera existed, and that the data from that camera would be accessible for as long as the program ran. This update gets rid of those assumptions by checking to see if the camera is being captured, and checking to see if the frame data is being recieved.