Don't detect faces from camera if this module is being imported
Previously, this file entered a while loop to process a video input no matter how this code was run. This meant that if you wanted to import `face_recognition` into your code, that import would cause this file to start capturing video. Now, this should make sure that this only happens when `face_recognition` is the file being run.